[FEATURE] Enable daemon mode
Create a daemon mode to go along with running a single execution.

Includes a FileWatcher to check for new files that match the configured
characteristics. Instead of doing these checks at the time of processing
it is handled up front to reduce overhead.

Relocated functionality between tools, episode, and formatter modules
based on proper location.